Add missing indices
A lot of indices through whole DB were missing. While some of them are
overlapping with existing mutlicolumn indices the order of columns is
wrong, as most specific ones are first and least specific (and by that
most used) are last which mean that DB cannot use them in queries.