Add sync type passive as a comparer in sync operation
The "sync_type" option can now be passed to the sync request to have a
different way to handle operations.

The new "passive" option will ignore all actions but the `new`, `renew`
and `remove`. It will ensure that when uploading a file, no translations
or master strings can be overwritten. It behaves the same way as the
existing merge_type "passive".