Add GitHub webhooks integration to sync a project (#85)
* Add GitHub webhooks integration to sync a project

The webhook generates events that reads the repo’s config file (the same
config file used by accent-cli) and build "sync" and "add translations"
events in Accent.

It supports a default ref: you can specify to only
trigger the "sync" on the "release/production" branch. It also supports
automatic version sync. If you have a version in Accent that matches
your git tag, "sync" events will be generated for those versions,
without touching your current translations. This can be useful if you
use versions to freeze translations on your releases.

This commit only adds the API side of the feature. So while it’s not
user facing, if you want to test the feature, you can add a matching
"integrations" row in the database and setup your webhook in GitHub :)

* Add tests for integration scope

* Rename hook sync to hook update since the hook sync and add translations

* Order mix.exs deps

* Run formatter

* Fix credo

* Fix tests

* Move GitHub producer with the same shared code as the other producer, simple queue without back pressure

* Remove unused function

* Update event producer from gen stage queue broadcaster

* Add tests