Fix CLI where files are being synced from the project document instead of from the filesystem
CLI sync should not create files as this is the role of the versionning
system (git). This allow a project to contain file from 2 different
repos (like a webapp and a API CMS).

This is a fix from a recent addition that was also consider a fix. The
fix was wrong so the document-paths-fetcher.ts file is patched.

The stats command also reflect which document from Accent’s project is
found in the filesystem.